project/requirements.txt
python-telegram-bot>=20.0,<21.0
aiosqlite>=0.18
pillow>=10.0
pytesseract>=0.3.10
opencv-python-headless>=4.8
numpy>=1.24
pytz>=2023.3
croniter>=2.0
pytest-asyncio>=0.23
